requirements.txt
pyserial==3.5
esptool==4.6.2
pyinstaller==6.3.0
